+++ changelog.md
@@ -18,6 +18,7 @@ Nov 24, 2022
 - Improving contrast for old items/ checked tasks
 - Raplacing name WeekToDo for WeekToDo
 - improving dark theme switchers btn
+- Adding details to recurrent events frecuency
 
 ## v1.9.0
 Jul 28, 2022